upadte ln-diag-initramfs-images.sh

+++ ln-diag-initramfs-images.sh
@@ -3,6 +3,8 @@
 # ln-diag-initramfs-images.sh
 #
 DIAG_INITRAMFS_NAME="diag-kchen-image-initramfs"
+DIAG_INITRAMFS_NAME="diag-image-initramfs"
+
 SRC_IMGAE="soniccorex-image-sunup-initramfs-${BUILD_MACHINE}.cpio.gz"
 DEST_IMAGE="${DIAG_INITRAMFS_NAME}-${BUILD_MACHINE}.cpio.gz"
 #
@@ -11,20 +13,26 @@ if [ "$1" = "" ] ; then
     printf "For example:\n\t ./ln-diag-initramfs.sh anderson-peak\n\n"
     return 1 
 fi 
-
+#
 pushd .
 source ~/sh-bitbake/cd-deploy-images.sh $1
+if [ $? -ne 0 ] ; then 
+    popd 
+    return 2 
+fi 
+#
 rm -f ${DEST_IMAGE}
 if [ $? -ne 0 ] ; then 
     echo -e "\nERROR: \"rm -f ${DEST_IMAGE}\n"
     popd 
-    return 2 
+    return 3 
 fi 
+#
 ln -s ${SRC_IMGAE} ${DEST_IMAGE}
 if [ $? -ne 0 ] ; then 
     echo -e "\nERROR: \"ln -s ${SRC_IMGAE} ${DEST_IMAGE}\" failed!\n"
     popd
-    return 3 
+    return 4 
 fi 
 #
 echo
